Control.Monad.Class.MonadSTM.Strict

Documentation

catchSTM ∷ ∀ (m ∷ Type → Type) e a. (MonadSTM m, MonadCatch (STM m), Exception e) ⇒ STM m a → (e → STM m a) → STM m a

catch speclialized for an stm monad.

throwSTM ∷ ∀ (m ∷ Type → Type) e a. (MonadSTM m, MonadThrow (STM m), Exception e) ⇒ e → STM m a

throwIO specialised to stm monad.

flushTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m [a]

lengthTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Natural

isFullTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

isEmptyTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

writeTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → a → STM m ()

tryPeekTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

peekTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryReadTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

readTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m a

newTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ Natural → STM m (TBQueueDefault m a)

labelTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TBQueueDefault m a → String → STM m ()

tryPeekTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

peekTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m a

isEmptyTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m Bool

tryReadTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

readTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m a

writeTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → a → STM m ()

newTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ STM m (TQueueDefault m a)

labelTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TQueueDefault m a → String → STM m ()

isEmptyTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m Bool

swapTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m a

tryReadTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

readTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m a

tryPutTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m Bool

putTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m ()

tryTakeTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

takeTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m a

newEmptyTMVarMDefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

newEmptyTMVarIODefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

newEmptyTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ STM m (TMVarDefault m a)

newTMVarMDefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newTMVarIODefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ a → STM m (TMVarDefault m a)

traceTMVarDefault ∷ ∀ (m ∷ Type → Type) tr proxy a. (MonadTraceSTM m, Typeable tr) ⇒ proxy m → TMVarDefault m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → STM m ()

labelTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TMVarDefault m a → String → STM m ()

type family TBQueue (m ∷ Type → Type) ∷ Type → Type

type family TQueue (m ∷ Type → Type) ∷ Type → Type

type family STM (m ∷ Type → Type) = (stm ∷ Type → Type) | stm → m

class (Monad m, Alternative (STM m), MonadPlus (STM m)) ⇒ MonadSTM (m ∷ Type → Type) where

Minimal complete definition

atomically, newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, writeTQueue, isEmptyTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue

Associated Types

type STM (m ∷ Type → Type) = (stm ∷ Type → Type) | stm → m

type TQueue (m ∷ Type → Type) ∷ Type → Type

type TBQueue (m ∷ Type → Type) ∷ Type → Type

Methods

atomically ∷ HasCallStack ⇒ STM m a → m a

retry ∷ STM m a

orElse ∷ STM m a → STM m a → STM m a

modifyTVar' ∷ TVar m a → (a → a) → STM m ()

check ∷ Bool → STM m ()

newTQueue ∷ STM m (TQueue m a)

readTQueue ∷ TQueue m a → STM m a

tryReadTQueue ∷ TQueue m a → STM m (Maybe a)

peekTQueue ∷ TQueue m a → STM m a

tryPeekTQueue ∷ TQueue m a → STM m (Maybe a)

writeTQueue ∷ TQueue m a → a → STM m ()

isEmptyTQueue ∷ TQueue m a → STM m Bool

newTBQueue ∷ Natural → STM m (TBQueue m a)

readTBQueue ∷ TBQueue m a → STM m a

tryReadTBQueue ∷ TBQueue m a → STM m (Maybe a)

peekTBQueue ∷ TBQueue m a → STM m a

tryPeekTBQueue ∷ TBQueue m a → STM m (Maybe a)

flushTBQueue ∷ TBQueue m a → STM m [a]

writeTBQueue ∷ TBQueue m a → a → STM m ()

lengthTBQueue ∷ TBQueue m a → STM m Natural

Since: io-classes-0.2.0.0

isEmptyTBQueue ∷ TBQueue m a → STM m Bool

isFullTBQueue ∷ TBQueue m a → STM m Bool

newTQueueIO ∷ m (TQueue m a)

newTBQueueIO ∷ Natural → m (TBQueue m a)

class MonadSTM m ⇒ MonadLabelledSTM (m ∷ Type → Type) where

Labelled TVars, TMVars, TQueues and TBQueues.

Minimal complete definition

labelTVar, labelTMVar, labelTQueue, labelTBQueue

Methods

labelTQueue ∷ TQueue m a → String → STM m ()

labelTBQueue ∷ TBQueue m a → String → STM m ()

labelTQueueIO ∷ TQueue m a → String → m ()

labelTBQueueIO ∷ TBQueue m a → String → m ()

type family InspectMonad (m ∷ Type → Type) ∷ Type → Type

class (MonadSTM m, Monad (InspectMonad m)) ⇒ MonadInspectSTM (m ∷ Type → Type) where

This type class is indented for 'io-sim', where one might want to access TVar in the underlying ST monad.

Associated Types

type InspectMonad (m ∷ Type → Type) ∷ Type → Type

Methods

inspectTVar ∷ proxy m → TVar m a → InspectMonad m a

inspectTMVar ∷ proxy m → TMVar m a → InspectMonad m (Maybe a)

class MonadInspectSTM m ⇒ MonadTraceSTM (m ∷ Type → Type) where

MonadTraceSTM allows to trace values of stm variables when stm transaction is commited. This allows to verify invariants when a variable is commited.

Minimal complete definition

traceTVar, traceTQueue, traceTBQueue

Methods

traceTVar

Arguments

∷ Typeable tr

⇒ proxy m

→ TVar m a

→ (Maybe a → a → InspectMonad m tr)
callback which receives initial value or Nothing (if it is a newly created TVar), and the commited value.

→ STM m ()

traceTMVar ∷ Typeable tr ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → STM m ()

traceTQueue ∷ Typeable tr ⇒ proxy m → TQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → STM m ()

traceTBQueue ∷ Typeable tr ⇒ proxy m → TBQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → STM m ()

traceTVarIO ∷ Typeable tr ⇒ proxy m → TVar m a → (Maybe a → a → InspectMonad m tr) → m ()

traceTMVarIO ∷ Typeable tr ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → m ()

traceTQueueIO ∷ Typeable tr ⇒ proxy m → TQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

traceTBQueueIO ∷ Typeable tr ⇒ proxy m → TBQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

data TMVarDefault (m ∷ Type → Type) a

data TQueueDefault (m ∷ Type → Type) a

Constructors

TQueue !(TVar m [a]) !(TVar m [a])

data TBQueueDefault (m ∷ Type → Type) a

Constructors

TBQueue !(TVar m Natural) !(TVar m [a]) !(TVar m Natural) !(TVar m [a]) !Natural

newtype WrappedSTM (t ∷ Trans) r (m ∷ Type → Type) a

A newtype wrapper for an STM monad for monad transformers.

Constructors

WrappedSTM

Fields

runWrappedSTM ∷ STM m a

type LazyTVar m = TVar m

type LazyTMVar m = TMVar m

StrictTVar

data StrictTVar m a

labelTVar ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → STM m ()

labelTVarIO ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → m ()

castStrictTVar ∷ LazyTVar m ~ LazyTVar n ⇒ StrictTVar m a → StrictTVar n a

toLazyTVar ∷ StrictTVar m a → LazyTVar m a

Get the underlying TVar

Since we obviously cannot guarantee that updates to this LazyTVar will be strict, this should be used with caution.

newTVar ∷ MonadSTM m ⇒ a → STM m (StrictTVar m a)

newTVarIO ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

newTVarWithInvariantIO

Arguments

∷ (MonadSTM m, HasCallStack)

⇒ (a → Maybe String)
Invariant (expect Nothing)

→ a

→ m (StrictTVar m a)

readTVar ∷ MonadSTM m ⇒ StrictTVar m a → STM m a

readTVarIO ∷ MonadSTM m ⇒ StrictTVar m a → m a

writeTVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictTVar m a → a → STM m ()

modifyTVar ∷ MonadSTM m ⇒ StrictTVar m a → (a → a) → STM m ()

stateTVar ∷ MonadSTM m ⇒ StrictTVar m s → (s → (a, s)) → STM m a

swapTVar ∷ MonadSTM m ⇒ StrictTVar m a → a → STM m a

StrictTMVar

data StrictTMVar m a

labelTMVar ∷ MonadLabelledSTM m ⇒ StrictTMVar m a → String → STM m ()

labelTMVarIO ∷ MonadLabelledSTM m ⇒ StrictTMVar m a → String → m ()

castStrictTMVar ∷ LazyTMVar m ~ LazyTMVar n ⇒ StrictTMVar m a → StrictTMVar n a

newTMVar ∷ MonadSTM m ⇒ a → STM m (StrictTMVar m a)

newTMVarIO ∷ MonadSTM m ⇒ a → m (StrictTMVar m a)

newEmptyTMVar ∷ MonadSTM m ⇒ STM m (StrictTMVar m a)

newEmptyTMVarIO ∷ MonadSTM m ⇒ m (StrictTMVar m a)

takeTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m a

tryTakeTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

putTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m ()

tryPutTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m Bool

readTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m a

tryReadTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

swapTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m a

isEmptyTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m Bool

Low-level API

checkInvariant ∷ HasCallStack ⇒ Maybe String → a → a

Check invariant (if enabled) before continuing

checkInvariant mErr x is equal to x if mErr == Nothing, and throws an error err if mErr == Just err.

This is exported so that other code that wants to conditionally check invariants can reuse the same logic, rather than having to introduce new per-package flags.

Deprecated API

updateTVar ∷ MonadSTM m ⇒ StrictTVar m s → (s → (a, s)) → STM m a

Deprecated: Use stateTVar

newTVarM ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

Deprecated: Use newTVarIO

newTVarWithInvariantM

Arguments

∷ (MonadSTM m, HasCallStack)

⇒ (a → Maybe String)
Invariant (expect Nothing)

→ a

→ m (StrictTVar m a)

Deprecated: Use newTVarWithInvariantIO

newTMVarM ∷ MonadSTM m ⇒ a → m (StrictTMVar m a)

Deprecated: Use newTVarIO

newEmptyTMVarM ∷ MonadSTM m ⇒ m (StrictTMVar m a)

Deprecated: Use newEmptyTMVarIO